+++ lib/presentation/admin/daftar_shop.dart
@@ -107,15 +107,15 @@ class _DaftarShopState extends State<DaftarShop> {
           title: Column(
             mainAxisAlignment: MainAxisAlignment.center,
             children: [
-              ClipOval(
-                child: SizedBox.fromSize(
-                  size: const Size.fromRadius(38), // Image radius
-                  child: Image.asset('assets/ftprofil.png', fit: BoxFit.cover),
-                ),
-              ),
+              // ClipOval(
+              //   child: SizedBox.fromSize(
+              //     size: const Size.fromRadius(38), // Image radius
+              //     child: Image.asset('assets/ftprofil.png', fit: BoxFit.cover),
+              //   ),
+              // ),
               Container(
                   padding: const EdgeInsets.all(8.0),
-                  child: const Text('Hello, Ivan'))
+                  child: const Text('Shop'))
             ],
           ),
           shape: const RoundedRectangleBorder(
@@ -211,13 +211,50 @@ class _DaftarShopState extends State<DaftarShop> {
                             Row(
                               children: [
                                 InkWell(
-                                  onTap: () {
-                                    // Implementasi untuk menghapus barang
-                                    deleteProduct(product['id']);
-                                  },
-                                  child: const Icon(Icons.delete,
-                                      color: Colors.red),
-                                ),
+                                onTap: () {
+                                  showDialog(
+                                    context: context,
+                                    builder: (BuildContext context) {
+                                      return AlertDialog(
+                                        title: const Text('Konfirmasi'),
+                                        content: const Text(
+                                            'Apakah Anda yakin ingin menghapus aktivitas ini?'),
+                                        actions: [
+                                          TextButton(
+                                            onPressed: () {
+                                              Navigator.of(context)
+                                                  .pop(); // Tutup dialog
+                                            },
+                                            child: const Text('Batal'),
+                                          ),
+                                          TextButton(
+                                            onPressed: () {
+                                              Navigator.of(context)
+                                                  .pop(); // Tutup dialog
+                                              deleteProduct(product['id']); // Panggil fungsi delete
+                                            },
+                                            child: const Text(
+                                              'Hapus',
+                                              style:
+                                                  TextStyle(color: Colors.red),
+                                            ),
+                                          ),
+                                        ],
+                                      );
+                                    },
+                                  );
+                                },
+                                child:
+                                    const Icon(Icons.delete, color: Colors.red),
+                              ),
+                                // InkWell(
+                                //   onTap: () {
+                                //     // Implementasi untuk menghapus barang
+                                //     deleteProduct(product['id']);
+                                //   },
+                                //   child: const Icon(Icons.delete,
+                                //       color: Colors.red),
+                                // ),
                                 const SizedBox(
                                     width:
                                         8), // Jarak antara ikon delete dan edit
